Adds clarifying text to the EIP-155 specification. (#2272)
As someone implementing stuff related to signing in this space I have had to come back and read this EIP many times to remind myself how to properly sign/recover a transaction post EIP-155.  Each time, I read this EIP and am left confused about what I'm actually supposed to do when signing.  After once again digging to figure out what I'm supposed to do I decided to edit this EIP to hopefully prevent future developers (including myself) from running into the same frustration.

This change is non-normative and only attempts to improve clarity of the process.  If a normative change was made, this means my understanding is flawed and corrections should be made before this PR is merged.

The major change is to include a section for how to generate a valid post-155 signature.  The original text only explained how to _validate_ a post-155 signature and it was left up to the reader to figure out how to generate one.

I chose to use **SHOULD** here because it offers the users additional security against replay attacks and generally should be preferred unless you have good reason to not sign this way.  This could be widened to **MAY** if that is desirable.  **MUST** would be a normative change so that was intentionally not chosen.

I have also made a couple grammatical and formatting changes to the original text while editing which I felt helped improve readability.